Consolidate Webhook and Event types in to Event
* Despite being identical payloads from the API, we had previously
  created separate Webhook and Event types with very similar type
  signatures. The key difference was that the Event data was not
  strongly typed to the response we would get from the API.
* This change migrates to a single Event type that has a more specific
  response signature.
* Additionally, we moved the webhook specific serializers into their
  respective domains instead of bundling them under webhook serializers.
  This makes them more re-usable from a common location.
