+++ src/common/interfaces/index.ts
@@ -1,3 +1,4 @@
+export * from './event.interface';
 export * from './get-options.interface';
 export * from './list.interface';
 export * from './post-options.interface';
