+++ src/directory-sync/interfaces/directory-group.interface.ts
@@ -19,3 +19,25 @@ export interface DirectoryGroupResponse {
   updated_at: string;
   raw_attributes: any;
 }
+
+export interface EventDirectoryGroup<TRawAttributes = any> {
+  id: string;
+  idpId: string;
+  directoryId: string;
+  organizationId: string | null;
+  name: string;
+  createdAt: string;
+  updatedAt: string;
+  rawAttributes: TRawAttributes;
+}
+
+export interface EventDirectoryGroupResponse<TRawAttributes = any> {
+  id: string;
+  idp_id: string;
+  directory_id: string;
+  organization_id: string | null;
+  name: string;
+  created_at: string;
+  updated_at: string;
+  raw_attributes: TRawAttributes;
+}
