+++ src/directory-sync/serializers/directory.serializer.ts
@@ -1,4 +1,9 @@
-import { Directory, DirectoryResponse } from '../interfaces';
+import {
+  Directory,
+  DirectoryResponse,
+  EventDirectory,
+  EventDirectoryResponse,
+} from '../interfaces';
 
 export const deserializeDirectory = (
   directory: DirectoryResponse,
@@ -11,6 +16,34 @@ export const deserializeDirectory = (
   organizationId: directory.organization_id,
   state: directory.state,
   type: directory.type,
+  createdAt: directory.created_at,
+  updatedAt: directory.updated_at,
+});
+
+export const deserializeEventDirectory = (
+  directory: EventDirectoryResponse,
+): EventDirectory => ({
+  object: directory.object,
+  id: directory.id,
+  externalKey: directory.external_key,
+  type: directory.type,
+  state: directory.state,
+  name: directory.name,
+  organizationId: directory.organization_id,
+  domains: directory.domains,
+  createdAt: directory.created_at,
+  updatedAt: directory.updated_at,
+});
+
+export const deserializeDeletedEventDirectory = (
+  directory: Omit<EventDirectoryResponse, 'domains' | 'external_key'>,
+): Omit<EventDirectory, 'domains' | 'externalKey'> => ({
+  object: directory.object,
+  id: directory.id,
+  type: directory.type,
+  state: directory.state,
+  name: directory.name,
+  organizationId: directory.organization_id,
   createdAt: directory.created_at,
   updatedAt: directory.updated_at,
 });
