+++ src/events/interfaces/index.ts
@@ -1,2 +1,1 @@
 export * from './list-events-options.interface';
-export * from './event.interface';
