--- src/events/serializers/event.serializer.ts
@@ -1,8 +1,0 @@
-import { Event, EventResponse } from '../interfaces';
-
-export const deserializeEvent = (event: EventResponse): Event => ({
-  id: event.id,
-  createdAt: event.created_at,
-  data: event.data,
-  event: event.event,
-});
