+++ src/index.ts
@@ -11,7 +11,6 @@ export * from './organizations/interfaces';
 export * from './passwordless/interfaces';
 export * from './portal/interfaces';
 export * from './sso/interfaces';
-export * from './webhooks/interfaces';
 export * from './users/interfaces';
 
 export { WorkOS };
