--- src/webhooks/interfaces/index.ts
@@ -1,3 +1,0 @@
-export * from './webhook.interface';
-export * from './webhook-directory.interface';
-export * from './webhook-directory-group.interface';
