--- src/webhooks/serializers/index.ts
@@ -1,3 +1,0 @@
-export * from './webhook-directory-group.serializer';
-export * from './webhook-directory-user.serializer';
-export * from './webhook-directory.serializer';
